backend/requirements.txt
# existing dependencies...
Django==5.2.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
PyJWT==2.9.0
python-jose==3.3.0
requests==2.32.3
pydantic==2.9.2
pydantic-settings==2.6.1
gunicorn==23.0.0
uvicorn==0.32.0
whitenoise==6.8.2
python-dotenv==1.0.1

# Redis client for django caching
django-redis==5.4.0
redis==5.0.1
celery==5.6.3
